+++ src/lxp/packet.rs
@@ -12,7 +12,7 @@ pub enum ReadInput {
 }
 
 // {{{ ReadInput1
-#[derive(Debug, Serialize, Nom, InfluxDbWriteable)]
+#[derive(Debug, Serialize, Nom)]
 #[nom(LittleEndian)]
 pub struct ReadInput1 {
     pub status: u16,
@@ -96,16 +96,13 @@ pub struct ReadInput1 {
     pub v_bus_2: f64,
 
     #[nom(Parse = "utils::current_time")]
-    #[serde(serialize_with = "UnixTime::serialize")]
     pub time: UnixTime,
     #[nom(Ignore)]
-    #[serde(skip)]
-    #[influxdb(tag)]
-    pub datalog: Serial, // serde skips this so its not in MQTT messages
+    pub datalog: Serial,
 } // }}}
 
 // {{{ ReadInput2
-#[derive(Debug, Serialize, Nom, InfluxDbWriteable)]
+#[derive(Debug, Serialize, Nom)]
 #[nom(Debug, LittleEndian)]
 pub struct ReadInput2 {
     #[nom(Ignore)]
@@ -143,16 +140,13 @@ pub struct ReadInput2 {
     // bunch of auto_test stuff here I'm not doing yet
     //
     #[nom(Parse = "utils::current_time")]
-    #[serde(serialize_with = "UnixTime::serialize")]
     pub time: UnixTime,
     #[nom(Ignore)]
-    #[serde(skip)]
-    #[influxdb(tag)]
-    pub datalog: Serial, // serde skips this so its not in MQTT messages
+    pub datalog: Serial,
 } // }}}
 
 // {{{ ReadInput3
-#[derive(Debug, Serialize, Nom, InfluxDbWriteable)]
+#[derive(Debug, Serialize, Nom)]
 #[nom(LittleEndian)]
 pub struct ReadInput3 {
     #[nom(SkipBefore(2))] // bat_brand, bat_com_type
@@ -182,13 +176,41 @@ pub struct ReadInput3 {
 
     // following are for influx capability only
     #[nom(Parse = "utils::current_time")]
-    #[serde(serialize_with = "UnixTime::serialize")]
     pub time: UnixTime,
     #[nom(Ignore)]
-    #[serde(skip)]
-    #[influxdb(tag)]
-    pub datalog: Serial, // serde skips this so its not in MQTT messages
+    pub datalog: Serial,
 } // }}}
+
+#[derive(Debug, Serialize)]
+pub struct ReadInputs {
+    #[serde(flatten)]
+    read_input_1: Option<ReadInput1>,
+    #[serde(flatten)]
+    read_input_2: Option<ReadInput2>,
+    #[serde(flatten)]
+    read_input_3: Option<ReadInput3>,
+}
+
+impl ReadInputs {
+    pub fn default() -> Self {
+        Self {
+            read_input_1: None,
+            read_input_2: None,
+            read_input_3: None,
+        }
+    }
+
+    pub fn set_read_input_1(&mut self, i: ReadInput1) {
+        self.read_input_1 = Some(i);
+    }
+    pub fn set_read_input_2(&mut self, i: ReadInput2) {
+        self.read_input_2 = Some(i);
+    }
+
+    pub fn set_read_input_3(&mut self, i: ReadInput3) {
+        self.read_input_3 = Some(i);
+    }
+}
 
 // {{{ TcpFunction
 #[derive(Clone, Copy, Debug, PartialEq, IntoPrimitive, TryFromPrimitive)]