+++ src/unixtime.rs
@@ -3,7 +3,7 @@
 // than good enough and should let Influx store it more efficiently.
 
 use chrono::{DateTime, Utc};
-use serde::Serializer;
+use serde::{Serialize, Serializer};
 
 #[derive(Debug)]
 pub struct UnixTime(DateTime<Utc>);
@@ -12,22 +12,15 @@ impl UnixTime {
     pub fn now() -> Self {
         Self(Utc::now())
     }
+}
 
-    // default chrono serialization uses RFC3339 with nanosecond precision..
-    // a bit overkill for our uses. clamp it to seconds.
-    pub fn serialize<S>(u: &UnixTime, serializer: S) -> Result<S::Ok, S::Error>
+// default chrono serialization uses RFC3339 with nanosecond precision..
+// a bit overkill for our uses. clamp it to seconds.
+impl Serialize for UnixTime {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
         S: Serializer,
     {
-        // previously used to send ISO8601 string. left for reference.
-        // serializer.serialize_str(&u.0.to_rfc3339_opts(chrono::SecondsFormat::Secs, true));
-
-        serializer.serialize_i64(u.0.timestamp())
-    }
-}
-
-impl From<UnixTime> for influxdb::Timestamp {
-    fn from(u: UnixTime) -> Self {
-        influxdb::Timestamp::Seconds(u.0.timestamp() as u128)
+        serializer.serialize_i64(self.0.timestamp())
     }
 }